refactor(tests): reorganiza testes da feature de usuários em pasta Users
- Move arquivos CreateUserHandlerTests e TestData para pasta Users
- Ajusta namespaces para refletirem a nova estrutura
- Mantém cobertura de testes com sucesso
