+++ tests/Ambev.DeveloperEvaluation.Unit/Application/Users/TestData/CreateUserHandlerTestData.cs
@@ -2,7 +2,7 @@ using Ambev.DeveloperEvaluation.Application.Users.CreateUser;
 using Ambev.DeveloperEvaluation.Domain.Enums;
 using Bogus;
 
-namespace Ambev.DeveloperEvaluation.Unit.Domain;
+namespace Ambev.DeveloperEvaluation.Unit.Application.Users.TestData;
 
 /// <summary>
 /// Provides methods for generating test data using the Bogus library.
